+++ lib/handlers/toast_handler.dart
@@ -1,3 +1,4 @@
+import 'package:catcher/core/application_profile_manager.dart';
 import 'package:catcher/model/platform_type.dart';
 import 'package:catcher/model/report.dart';
 import 'package:catcher/model/report_handler.dart';
@@ -23,16 +24,41 @@ class ToastHandler extends ReportHandler {
     this.customMessage,
   });
 
+  FToast? fToast;
+
   @override
-  Future<bool> handle(Report error) async {
-    Fluttertoast.showToast(
-        msg: _getErrorMessage(error),
-        toastLength: _getLength(),
-        gravity: _getGravity(),
-        timeInSecForIosWeb: _getLengthIos(),
-        backgroundColor: backgroundColor,
-        textColor: textColor,
-        fontSize: textSize);
+  Future<bool> handle(Report error, BuildContext? buildContext) async {
+    if (ApplicationProfileManager.isAndroid() ||
+        ApplicationProfileManager.isIos() ||
+        ApplicationProfileManager.isWeb()) {
+      Fluttertoast.showToast(
+          msg: _getErrorMessage(error),
+          toastLength: _getLength(),
+          gravity: _getGravity(),
+          timeInSecForIosWeb: _getLengthIos(),
+          backgroundColor: backgroundColor,
+          textColor: textColor,
+          fontSize: textSize);
+    } else {
+      Future.delayed(
+        const Duration(milliseconds: 500),
+        () {
+          Navigator.push<void>(
+            buildContext!,
+            PageRouteBuilder(
+              opaque: false,
+              pageBuilder: (_, __, ___) => FlutterToastPage(
+                  _getErrorMessage(error),
+                  _getGravity(),
+                  Duration(seconds: _getLengthIos()),
+                  backgroundColor,
+                  textColor,
+                  textSize),
+            ),
+          );
+        },
+      );
+    }
 
     return true;
   }
@@ -78,4 +104,66 @@ class ToastHandler extends ReportHandler {
         PlatformType.android,
         PlatformType.iOS,
       ];
+
+  @override
+  bool isContextRequired() {
+    return true;
+  }
+}
+
+class FlutterToastPage extends StatefulWidget {
+  final String text;
+  final ToastGravity gravity;
+  final Duration duration;
+  final Color backgroundColor;
+  final Color textColor;
+  final double textSize;
+
+  const FlutterToastPage(this.text, this.gravity, this.duration,
+      this.backgroundColor, this.textColor, this.textSize,
+      {Key? key})
+      : super(key: key);
+
+  @override
+  _FlutterToastPageState createState() => _FlutterToastPageState();
+}
+
+class _FlutterToastPageState extends State<FlutterToastPage> {
+  FToast fToast = FToast();
+
+  @override
+  void initState() {
+    fToast.init(context);
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
+      showToast();
+    });
+    super.initState();
+  }
+
+  void showToast() {
+    fToast.showToast(
+        child: Container(
+          color: widget.backgroundColor,
+          child: Text(
+            widget.text,
+            style: TextStyle(
+              color: widget.textColor,
+              fontSize: widget.textSize,
+            ),
+          ),
+        ),
+        gravity: widget.gravity,
+        toastDuration: widget.duration);
+    Future.delayed(
+      Duration(milliseconds: widget.duration.inMilliseconds + 100),
+      () {
+        Navigator.of(context).pop();
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return const SizedBox();
+  }
 }